// src/_common/ts/env/enums.ts
// ******************************
// Enums:
// ******************************

export enum ENV_TYPE {
    Development = 'Development',
    Test = 'Test',
    Production = 'Production',
}

// ******************************

export enum LOG_LEVEL {
    Off = 0,
    Error = 1,
    Warning = 2,
    Info = 3,
    Verbose = 4,
    Debug = 5,
}

// ******************************

export enum API_METHOD {
    Post = 'POST',
    Get = 'GET',
    Put = 'PUT',
    Delete = 'DELETE',
}

// ******************************